+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 import './App.scss';
 import { withFirebase } from './Components/Firebase';
 import Header from './Components/Header/Header';
@@ -7,7 +7,7 @@ import * as ROUTES from "./Constants/routes";
 import Infrastructure from "./Pages/Infrastructure/Infrastructure";
 import Projects from "./Pages/Projects/Projects";
 import Settings from "./Pages/Settings/Settings";
-import Workspaces from "./Pages/Workspaces/Workspaces";
+import HomeWorkspaces from "./Pages/Workspaces/Index";
 import { compose } from 'recompose';
 import { connect } from 'react-redux';
 import { login, logout } from "./Store/actions/user.actions";
@@ -51,8 +51,8 @@ class App extends Component {
           <PrivateRoute exact isLoaded={userState.isLoaded} isAuthenticated={userState.isAuthenticated} path={ROUTES.PROJECTS} component={Projects} />
           <PrivateRoute exact isLoaded={userState.isLoaded} isAuthenticated={userState.isAuthenticated} path={ROUTES.INFRASTRUCTURE} component={Infrastructure} />
           <PrivateRoute exact isLoaded={userState.isLoaded} isAuthenticated={userState.isAuthenticated} path={ROUTES.SETTINGS} component={Settings} />
-          <PrivateRoute exact isLoaded={userState.isLoaded} isAuthenticated={userState.isAuthenticated} path={ROUTES.WORKSPACES} component={Workspaces} />
-          <Route exact path={ROUTES.LANDING} component={Landing} />
+          <PrivateRoute isLoaded={userState.isLoaded} isAuthenticated={userState.isAuthenticated} path={ROUTES.WORKSPACES} component={HomeWorkspaces} />
+          <PrivateRoute exact  isLoaded={userState.isLoaded} isAuthenticated={userState.isAuthenticated} path={ROUTES.LANDING} component={Landing} />
         </Router>
       </div>
     );
