+++ src/Constants/routes.js
@@ -5,4 +5,4 @@ export const INFRASTRUCTURE = '/infrastructure';
 export const PROJECTS = '/projects';
 export const SETTINGS = '/settings';
 export const PASSWORD_FORGET = '/pw-forget';
-export const WORKSPACES = '/workspaces';
+export const WORKSPACES = '/workspaces/:workspaceid';
