+++ src/Pages/Workspaces/Pages/Dashboard/Dashboard.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+
+class WorkspacesDashboard extends Component {
+    render(){
+        return <>
+        <h1>DASHBOARD FOR {this.props.match.params.workspaceid}</h1>
+     </>
+    }
+}
+export default WorkspacesDashboard
